backend/requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
anyio==4.4.0
asgi-lifespan==2.1.0
bcrypt==4.2.0
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
environs==11.0.0
exceptiongroup==1.2.2
fastapi==0.112.0
fastapi-mail==1.4.1
fastapi_cors==0.0.6
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
jose==1.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mongomock==4.2.0.post1
motor==3.3.2
packaging==24.1
passlib==1.7.4
pillow==10.4.0
pluggy==1.5.0
pyasn1==0.6.0
pydantic==2.8.2
pydantic-settings==2.5.0
pydantic_core==2.20.1
pymongo==4.6.2
pyotp==2.9.0
pytest==8.3.3
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.2
reportlab==4.2.2
requests==2.32.3
rsa==4.9
sentinels==1.0.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5